# docs/4_openapi.yaml
openapi: 3.1.0
info:
  title: 남이 차린 밥상 (Nam-Bapsang) Live Production API Specification
  description: 'SKALA Full-stack Engineering 웹 서비스 개발 Mini Project (지도강사: 권기창 강사님)

    Render 클라우드 실서버에서 100% 구동 중인 실체 API (REST & Realtime WebSocket) 규격서입니다.'
  version: 1.0.0
  contact:
    name: 박세웅 (Full-Stack Engineer)
    email: master@bapsang.com
paths:
  /api/v1/auth/signup:
    post:
      tags:
      - Auth & Profile
      summary: Signup
      description: 신규 회원가입 및 PostgreSQL DB 저장 (단 1개의 마스터 계정만 허용)
      operationId: signup_api_v1_auth_signup_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/login:
    post:
      tags:
      - Auth & Profile
      summary: Login
      description: '로그인 및 JWT Access Token 발급 (기본 마스터 계정: master@bapsang.com / master123)'
      operationId: login_api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/me:
    get:
      tags:
      - Auth & Profile
      summary: Get My Profile
      description: 내 프로필 정보 조회
      operationId: get_my_profile_api_v1_auth_me_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    LoginRequest:
      properties:
        email:
          type: string
          title: Email
        password:
          type: string
          title: Password
      type: object
      required:
      - email
      - password
      title: LoginRequest
    SignupRequest:
      properties:
        email:
          type: string
          title: Email
        password:
          type: string
          title: Password
        nickname:
          type: string
          title: Nickname
        master_key:
          anyOf:
          - type: string
          - type: 'null'
          title: Master Key
      type: object
      required:
      - email
      - password
      - nickname
      title: SignupRequest
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
servers:
- url: https://nam-bapsang-backend.onrender.com
  description: Render Production Cloud Server (실서버)
- url: http://localhost:8000
  description: Local Development Server (개발 서버)